examples/python-02-simple-chatbot/pyproject.toml
[tool.poetry]
name = "assistant"
version = "0.1.0"
description = "Exploration of a python Semantic Workbench OpenAI assistant."
authors = ["Semantic Workbench Team"]
readme = "README.md"
packages = [{ include = "assistant" }]

[tool.poetry.dependencies]
python = "~3.11"
aiohttp = "^3.10.5"
azure-ai-contentsafety = "^1.0.0"
azure-core = "^1.31.0"
azure-identity = "^1.17.1"
openai = "^1.3.9"
tiktoken = "^0.7.0"
# If you copy this file to your project, you should verify the relative path to the following:
semantic-workbench-assistant = { path = "../../semantic-workbench/v1/service/semantic-workbench-assistant", develop = true }

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.4"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
exclude = ["venv", ".venv"]
